Created register token tests in new class in other project. Removed old integration test project since migration complete.
