+++ PasswordlessServer.sln
@@ -7,8 +7,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Service", "src\Service\Serv
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Api", "src\Api\Api.csproj", "{67940251-EED0-4D8B-A733-51C893390A65}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Api.IntegrationTests", "tests\Api.IntegrationTests\Api.IntegrationTests.csproj", "{3C8BC355-5D8D-4FDA-AD17-2FC163B92A4C}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AdminConsole", "src\AdminConsole\AdminConsole.csproj", "{19FE0682-DA56-49F5-8C04-ECA06CC08F4B}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{A5E8A3D4-EA0F-42EF-B78C-4340E430DF42}"
@@ -54,10 +52,6 @@ Global
 		{67940251-EED0-4D8B-A733-51C893390A65}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{67940251-EED0-4D8B-A733-51C893390A65}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{67940251-EED0-4D8B-A733-51C893390A65}.Release|Any CPU.Build.0 = Release|Any CPU
-		{3C8BC355-5D8D-4FDA-AD17-2FC163B92A4C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{3C8BC355-5D8D-4FDA-AD17-2FC163B92A4C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{3C8BC355-5D8D-4FDA-AD17-2FC163B92A4C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{3C8BC355-5D8D-4FDA-AD17-2FC163B92A4C}.Release|Any CPU.Build.0 = Release|Any CPU
 		{19FE0682-DA56-49F5-8C04-ECA06CC08F4B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{19FE0682-DA56-49F5-8C04-ECA06CC08F4B}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{19FE0682-DA56-49F5-8C04-ECA06CC08F4B}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -96,7 +90,6 @@ Global
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{C38C006F-89B8-4076-BFFC-F0AA99FB8560} = {A5E8A3D4-EA0F-42EF-B78C-4340E430DF42}
-		{3C8BC355-5D8D-4FDA-AD17-2FC163B92A4C} = {A5E8A3D4-EA0F-42EF-B78C-4340E430DF42}
 		{A5969EF2-0466-400F-87D4-DB56E2193566} = {A5E8A3D4-EA0F-42EF-B78C-4340E430DF42}
 		{E77FC7F6-3324-4CA4-8354-A04E0DE8C74A} = {A5E8A3D4-EA0F-42EF-B78C-4340E430DF42}
 		{B2FB89E3-F867-44AF-B643-D213D64D45D4} = {A5E8A3D4-EA0F-42EF-B78C-4340E430DF42}
